Initial work on migrating some of the rules in initnull's Tachyon Web Discovery system to Vega.
Up to my knowledge and understanding of the code, this module should enqueue
extra checks, however it is not, it seems ctx.submitRequest calls are being ignored

It is unclear whether a module license is important regarding EPL vs. GPL. In the
worse case we can ask initnull's to re-license these files.